Build: allows to disable www setup
The set_www_env function is used to setup the web environment for
test. It requires sudo permissions in order to modify the web
server/PHP config and restart it.
But in some case (like for instance when using a container image),
it's not needed. The system already have the correct configuration and
we do not need this setup.
This patch allows to skip the apache config step if the
SKIP_HTTPD_SETUP environment variable is set.

Change-Id: I9b37cccea5ae6e7870d768d844292afdf00940bb