core/obj: trace refs created by 'obj_retain_scope()'
Include the object references created by 'obj_retain_scope' in
the data structures allowing us to trace tagged references.

There cannot be balance issues with references created with
'obj_retain_scope' but it can help to be able to distinguish these
references from untagged ones created with 'obj_retain()' when printing
the references with 'obj_print_references()'.

With this patch, the references are traced just like tagged references,
but there is no need to set tags for them: they are identified with
their file and line.

Change-Id: I5c61a25362a58015a00807d154b9dd8cca8efc7a