core/obj: introduce tagged references
Reference counting can be messy. In some code paths, it can be
absolutely not obvious to match the 'obj_retain()' associated to each
'obj_release()', making the code really hard to maintain.

This patch introduced tagged references. A tagged reference is created
with an 'obj_tagged_retain()' and can only be release with an
'obj_tagged_release()' using the same tag:

 obj_tagged_retain(my_obj, my_tag);
 /* Do stuff. */
 obj_tagged_release(&my_obj, my_tag);

This should be used as a tool for non-obvious use cases. It can also be
useful when fixing reference counting in some messy code.

Note: 'obj_tagged_retain()' and 'obj_tagged_release()' only exist in
debug mode. In release mode they are respectively replaced by
'obj_retain()' and 'obj_release()'.

Change-Id: Ieb00151e757a08e04e646a3bc207419efdb781c6