waftools/backend: fix copying stlib tgen for fpic.
When compiling the static libraries with '-fPIC' for shared libraries,
we are making a copy of the stlib TaskGen to compile all the source
files with '-fPIC' to a different target.

Unfortunately, not all the attributes of the original stlib TaskGen were
copied, only a subset of these attributes.

This caused issues when the original TaskGen used attributes that were
not listed, the attributes were simply just not copied.

The solution is to not rely on a list of copied attributes, but to copy
all the attributes of the original TaskGen while skipping some internal
attributes.

TaskGen.clone() does not work in our case because it does not create a
stlib TaskGen, but a generic TaskGen. Moreover, it copies some
attributes that should not be copied.

Change-Id: Id2a4c8ff93ae7fa6bb833d6a8a6e3fc4dc8a8077