lib-common/zchk: new function to parse human readable hexa strings
the 't_z_hex_decode()' function is often used in zchk tests when a test
uses a payload but it is redefined every time.

This patch defines the function 't_lstr_human_hexdecode()' with the same
result and makes it possible to remove every definitions of the
't_z_hex_decode()' function and keep only one in the lib-common.

In addition to handle hexadecimal strings with spaces, the
't_lstr_human_hexdecode()' will exclude every non-hexadecimal
characters.

Refs: #80170
Change-Id: I22028d2923c7461b76730056b72eeb7081a0a8a9