iop/iop-core-obj: avoid crashing when getting an unregistered class
The generated function "<cls_pfx>_get_cls()" included a "cls_cast" on
the returned pointer. "cls_cast" crashes if passed a null pointer, so if
we were trying to get the class of an unregistered IOP core obj, we got
a crash that is quite hard to diagnose because it's hidden in some
generated code. Make the function support null values instead.

Change-Id: I837d83a09eb220efc45577e811344b4a3bfa9da8