qps: Enhance snapshot start/end logs.
These logs were emitted on the tracing logger at debug level, so were
not emitted by default. But they bring some important information, and a
QPS snapshot is a rare event so they are not floody. Emit them as debug
on the main logger so that they are available by default.

Also:
* use t_time_spent_to_str to print the spent time.
* do not print the qps pointer, this is useless information, especially
  now that we have one logger per QPS instance, with a proper name.

Refs: #87886
Change-Id: Ia0e273c5f1e5fb8197ec7df5d60108bd810ea81c