core-error: Do not dump memory maps by default in .debug files.
Before this patch, the memory maps were dumped by default in the .debug
files generated in case of crash.

But by experience, this information was almost never used for debugging,
and it really makes the .debug files "big". For example, I have checked
a debug file attached to a random ticket on out bug tracker, and the
file took 2 MB, and only 16 KB without the maps.

Having big debug files basically prevents from writing a automatic tool
to get them all from a production platform to attach them to tickets,
which is a shame.

This commit disables the maps by default, that can still be activated
using the IS_DEBUG_FILES_DUMP_MAPS environment variable.

Change-Id: Iaa4de5a7a18752755378fd8968160b51ea52e5d5