waftools: create iop_compile_commands.json
In order to have a better integration for IOP in tools, this patch
allows the creation of a iop_compile_commands.json file.
This file works just like the clangd compile_commands.json. For each
file, the compilation command is added to the file.
This will allows tools to have the correct context to handle an IOP file.

Change-Id: Ibc237d5a97b4a0663e3b8018d209e7e5bf35fcea