str-outbuf: add documentation for ob_add_memchunk
Added documentation for the function ob_add_memchunk() with a notice to
prevent misuse. In the past, we used it instead of of ob_add() without
noticing that one must guarantee that the memory block must outlive the
outbuf_t object when the input parameter `is_const` is true. Potential
use-after-free bugs were masked by the fact that ob_add_memchunk is
optimized to do ob_add() for chunk smaller than OUTBUF_CHUNK_MIN_SIZE
(which is currently equal to 16 KiB).

Change-Id: Ic9634a13c5fbc2163bccd533d1f3cca2ee8f8392