net/http: add missing break
A missing break in the switch on httpd->state can make the
http2_stream_on_data_server function try to parse a chunk's header after
parsing a http body.

This commit adds a break to avoid trying to parse a chunk's header after
trying to parse the body (composed of said chunks).

Change-Id: Ieea7fe7aac453b7fc584a897fdb5e36532aa3851