asdf: force lib prefix in python2 link flags
Unless you have the patched version of Debian of the python2-config
utility the path towards the Python 2 libraries is not provided when
asking the ldflags.

It means that the link cannot be done if python2 isn't installed in
a standard path (which is the case when using ASDF). So we add this path
manually which should work for all distribs.

Change-Id: I67c0eb346dfd7c4f131bdcbca68f590345efd2b3