z: Add Z_ASSERT_DBL_EQ.
Check for double equality with a simple == does not always work even
when the two values seem to be equal. For example, here is the kind of
output we could have using Z_ASSERT_EQ on two doubles:

 : assertion failed: distance == 30988.658931
 :  30988.658931 (distance)
 :  30988.658931 (30988.658931)

This patch adds Z_ASSERT_DBL_EQ, that takes two doubles and a precision,
and checks they are equal with this precision.

Change-Id: I1f81a3138d585b16089859a73408f1620d510ea5