waf[compilation_database]: Remove useless and buggy call to pre_build.
This code overloads the 'build' command to produce the files
compile_commands.json and iop_compile_commands.json at each build.

It was calling self.pre_build(), that runs the pre functions (registered
with ctx.add_pre_fun. The problem is that the standard implementation of
the build is also called, so the pre functions were called twice when
running the build command (and not the other commands, like 'check').

In particular, the pre function add_scan_in_signature (cf. fc0e4d9a56),
that adds the scan function in the tasks signature, was called twice
when building. And because when running 'waf check' this is not the
case, the signature of the tasks having a scan method was not the same
when running 'waf' or 'waf check'. As a consequence, running 'waf check'
after a compilation re-compiled a lot of stuff.

To fix that, just remove the pre_build call. After some tests, it seems
that the produced files are the sames with and without this call.

Change-Id: I3124db570c045ee87f8b895e09f6bbff21ceb4e6