net/http: Force sending client connection preamble.
Make sure that the http2 client sends the connection preamble once the
tcp connection is started, currently the client waits for the server to
send its preamble which is its initial setting frame. However this can
result in deadlock for servers (like tornado_http2 we've been using in
our case) that wait for clients to send their connection preamble before
doing the same. Standard does not stipulate that the server should wait
to receive the client connection preface. Anyway, it was bug on our
side.

Refs: #94454
Change-Id: I143b76eaae79d7d1b7204e6c2c6e887f2ae81717