core-error: Warn at the beginning of .debug when user context available.
When some additional user context is available (cf. debug_stack_scope),
it is dumped at the end of the file. The reason is that it calls user
code that already crashed to print the information, so it might crash
again and we don't want to truncate the debug file when it happens.

The problem is that most of time, people do not think about checking the
end of the file, and can miss some precious information that are
available for debug.

So when the debug stack is not empty, add a "warning" message before the
backtrace to clearly indicate there is something interesting to read at
the end.

Change-Id: I818be4922a8e4792f8e1c9f1895f7de55472c60b