qps/hat: clean "qhat_check_consistency()" return type
The function "qhat_check_consistency()" used to return a boolean: 'true'
if everything was OK, 'false' in case of suboptimality in the QHAT.

Of course, this is totally out of line with our coding rules.

In case of critical issue in the QHAT, the check ended with a "panic",
explaining the absence of need for other return values.

As I'm planning to add some features to the function, we're going to
need a proper return type.

This patch is not iso-functional as we will not emit some intermediate
warnings we used to display but it does not really matter AFAICT.

Refs: #94699
Change-Id: Ia1723546053bce92559d2405677961dd7a65aab5