qps/hat: fix missing suboptimality detection
The function 'qhat_node_check_consistency()' detects critical issues in
a QHAT and also suboptimalities: when the QHAT is not broken but is
not organized as we expected it to be.

This patch fixes a case when the detection of a suboptimality by
'qhat_node_check_child()' is ignored by the caller.

Refs: #94699
Change-Id: Ia2d12e38631f49291f09b875da8b4e4812e5358c