qps/hat: fix code homogeneity issue with PATH_STRUCTURE_CHANGED
When the structure of a QHAT is modified (nodes inserted/removed, leafs
flatten/unflatten, etc), the struct_gen of the hat is incremented with
the macro PATH_STRUCTURE_CHANGED().

There were something a little odd with this around leaf
flattening/unflattening: the 'unflatten' function in 'qhat.in.c' called
the macro but the 'flatten' function did not: it relied on the caller
instead. Always do it in the 'qhat.in.c' function to improve code
homogeneity.

Refs: #94699
Change-Id: Ib7398bd2a3dfe45866fea32f9e92b351c842c796