qps/hat: fix bad return value in 'remove' function
When removing an entry from a QHAT, the function 'qhat_remove()' returns
'true' if an entry was actually removed and 'false' if there were no
entry for the provided key.

This return value could be wrong for absent values if the removed value
was in a flatten leaf: we always returned 'true' while the value could
be already null (== missing).

This fix is not expected to have any impact though, at least of the QHAT
usage we know.

Refs: #94699
Change-Id: I0bfa2a129f5ba0ee148c1f353ff4f918af87eb7d