pxcc: fix elaborated type const with llvm16.
With llvm16 (CINDEX_VERSION >= (0, 63)), the resolved elaborated named
type is not const like it is with previous versions.

So 'const' was not printed when printing the final resolved type.

This is an issue when using a pointer to const values.
The following type:
    const int *tab;
Was printed as:
    int *tab;

The solution is before resolving the elaborated type, check if it is
'const', and force it after resolving the final type.

In t_print_field_type(), we now need to be able to break out of the loop
without returning directly from the function with a new variable
'end_loop'.

Change-Id: If7c30389bea230712a7caae1a6d1648263b169c7
Closes: #95397