qps/hat: implement a closer tracking for QHAT changes
We should always use the safe version of enumerator functions when the
QHAT is modified during the lifetime of the enumerator, otherwise we may
get broken data or even corrupt the QHAT.

When using "unsafe" QHAT functions, we have assertions that check that
we should not be using the "safe" version. Example:

 assert(en->path.generation == en->path.hat->struct_gen);

The assert above is triggered if the enumerator is out of sync with the
QHAT, meaning that changes have been done in the QHAT since the last
time we used the enumerator.

Problem: the generation changes only when the tree structure changes,
but that does not happen for every insertion/removal in the QHAT.
As a consequence, we may fail to detect misuses.

This patch adds a "hat_gen" that is incremented for each insertion or
removal attempt, even if they have no effect. It should help us detect
more easily when an "unsafe" access should have been "safe".

Refs: #94699
Change-Id: Iee0f8e8cd7f9d2d2e30ebd4d3eb022528e38103d