qps/hat: add a function that checks if a QHAT value is zero
We had several ways to check if a QHAT value was zero or not, this patch
adds a function so that we always check it the same way. It has its
importance because zero == absent for values stored in a flat leaf for
non nullable QHATs.

Also remove 'SET_ZERO*" unused macros.

Change-Id: Id7f91768e4c3461aea5254850c574e7b080f090b