z: improve the way we print IOP structs when comparison fails
In case of test failures while using testing macros like
Z_ASSERT_IOPEQUAL or similar IOP equality checks, the IOPs were printed
in JSON with IOP_JPACK_MINIMAL flags, which was basically unreadable by
a human being except when the IOP struct was small.

In order to fix that, this patch changes the encoding to multi-line
YAML and uses diff to emphasize the differences.

How it looked before (while voluntarily breaking an IOP test):

: tests/zchk-iop.c:2935: assertion failed: &obj_struct == &_exp
:  field `b.b.a`: value differs (`2` vs `3`)
:  &obj_struct:
:   {"a":0,"b":{"a":1,"b":{"a":2}},"c":[{"a":4},{"a":2}]}
:  &_exp:
:   {"a":0,"b":{"a":1,"b":{"a":3}},"c":[{"a":4},{"a":2}]}

How it looks after this patch:

: tests/zchk-iop.c:2935: assertion failed: &obj_struct == &_exp
:  field `b.b.a`: value differs (`2` vs `3`)
:
: --- /home/foo/dev/lib-common/z.tst.14222.auBRq0/diff-QHp6kX/left	2023-07-18 18:03:43.000000000 +0200
: +++ /home/foo/dev/lib-common/z.tst.14222.auBRq0/diff-QHp6kX/right	2023-07-18 18:03:43.000000000 +0200
: @@ -2,7 +2,7 @@
:  b:
:    a: 1
:    b:
: -    a: 2
: +    a: 3
:  c:
:    - a: 4
:    - a: 2

Change-Id: I1fb6b040210dc529bb2b6549e0ad846a8d636bda