qps-hat/zchk: prepare NR test for enumerator issues
When getting the current enumerated value with
'qhat_enumerator_get_value_safe()' after it was removed, the behavior is
not consistent: depending on the context, it can return a NULL value
(which is good because the value was removed) or position the enumerator
on the next value and return it instead.

The documentation does not say anything about the expected behavior when
the current element is removed. Actually, there are three possible ones:

(1) undefined behavior

    => Current behavior, not acceptable IMHO because there is no
       evidence that it cannot happen in existing QHAT use cases.

(2) go to the next element

    => Bad idea as it can make us easily skip an element.

(3) return a null entry and keep the enumerator key as it is

    => Better option because this is the current behavior in some cases.

There is another issue when using 'qhat_tree_enumerator_go_to()' to go
to the current key after it was removed: in some cases, it stays on the
removed key while the documentation says that it should go to the next
key.

This patch prepares dedicated tests that reproduce the aforementioned
issues. FIXME's were left for parts that should be updated after the
fixes.

Change-Id: I83498cfbe9db9be381399a1d2035fb740cba3ff0