qps/hat: fix behavior inconsistency with 'get_value()'
When using 'qhat_tree_enumerator_get_value()' after the current entry
was removed, we could face two different behavior:

1. The removed entry was stored in a flat leaf. In that case, the
   enumerator stayed positioned on the missing entry.

2. The removed entry was stored in a compact leaf. In that case, the
   enumerator moved to the next entry.

The point 2 is problematic because it could make us miss an entry quite
easily.

Example:

 qhat_for_each_unsafe(en, &trie) {
     const void *v;

     if (...) {
         /* Remove the current key. */
         qhat_remove(&trie, en.key, NULL);
     }

     v = qhat_enumerator_get_value_safe(&en);
     /* Do things with the value. */
 }

Depending on the situation the last key was (in a flat or in a compact),
we may or may not skip the key after the removed one, which is probably
*not* what we expect.

This patch fixes it by detecting when we are in situation 2 and forces
the enumerator to stay on the current key after the 'get_value()', even
when the key was removed.

Change-Id: I767ec7993abdce643e0effeab2f4ef509c8721ce