iop/fieldpath: fix invalid indirection when targeting array of classes.
When targeting a class field, we always inserted an indirection step
after the field step.

However, this is not valid when targeting directly an array of classes
and not elements of the array.

This caused an invalid indirection and resulted on a segfault as seen
in #96239.

Granted, this is a weird use-case, but it is supported and we should not
crash anyway.

Also add a small test for scalar array sorting.

Closes: #96239
Change-Id: Ib6d1a52bd0c06e1d4c225040c30723387590b918