net/http: http2: fix a wrong assert on idle httpd
Fix a wrong assertion in http2_conn_assert_idle_httpd_invariants() about
an idle http2-transported httpd. We have workflows where we send the
query reply asynchronously to its on_done callback, IoW, the callback
schedules the reply to be sent later. So, in this case, an idle
http2-transported can have a query that is parsed but not yet answered
contrary to the assert in http2_conn_assert_idle_httpd_invariants().
Note that !q->hdrs_done implies !q->answered (which is an invariant from
the http1.x layer).

Refs: #96524
Change-Id: If44110dc626f920a7acaeb52cc01873565851da0