net/http: http2: fix handling async query reply
Fixed a bug for http2-transported httpd in workflows where we send the
query reply asynchronously to its on_done callback. In this case, the
query on_done callback schedules the reply to be sent later. When the
reply is written in such scenario, there is a possibility that we miss
this event if the http2 connection is idle since we forgot to signal it
to the undelying http2 layer. This patch corrects the bug by reporting
the http1.x httpd_set_mask event to the http2 layer.

RunTests: http2
Closes: #96524
Cc-Redmine: yes
Change-Id: I7b90e9f699f04f4015e899dcfd8a4d56317d3c7b