qps/hat: rename 'qhat_enumerator_get_value()' (add "_unsafe" suffix)
That function was implicitly unsafe (== it could mess up if the QHAT
associated to the enumerator was modified since last enumerator use).

Unsafe operations cannot be used in every situations and they should be
used knowingly, so this patch adds a "_unsafe" suffix to this function
name.

Also fix the outdated documentation for the related functions and rename
'qhat_enumerator_get_value_common()' as 'qhat_enumerator_get_value()' as
the name became vacant.

Change-Id: I8b52960e7a74d2e587db3940287ed15870a6674c