qps/hat: fix corruption case in QHATs
The QPS hats library includes the struct 'qhat_path_t'. That structure
contains the path to a given element of the QHAT. It is useful to spare
a lookup towards an element we want to set or remove when we already
have a enumerator pointing on it.

The QHAT enumerator always contain a 'qhat_path_t', which is private and
supposedly valid when the enumerator did not reach its end yet.

Problem: in 'qhat_enumerator_get_path()', we did not always check that
the enumerator was not done before setting the path. That could result
in returning an half-set path that would break the qhat when used.

The problem was not so easy to track because most of the time, the
returned path was valid even if the enumerator was done, it probably did
not point at the intended place though (*cough* *cough*).

Refs: #94699
Change-Id: I916c44b509a78cfdaf92b910b6a15d21939009b8