zcore/zunittest: fix invalid flags applied to test functions.
Since 59eb6a1f49da, we retrieve and join the different flags for a
single test function using a global '_ALL_FLAGS' dictionary.
The key of the dictionary is the id of the function (id(func)).

Here is the tricky part:

The value returned by the function id() is the memory value of the C
pointer for the Python value on CPython. It is unique for the lifetime
of Python value, as long as the Python value is still alive.

However, in the case of the decorators, wrappers, and class methods,
getting a class method and using it in a decorator actually create a new
Python value which is a ref to the underlying class method.

So the lifetime of the 'func' value in the wrapper is only the wrapper
function itself, and the Python value is deleted at then end of the
wrapper. This means that the memory get released and can be reused by
other Python objects.

This is actually what is happening, and we have multiple 'func' with the
same id, and thus the same key in the global '_ALL_FLAGS' dictionary.

And since we merge the different for the same 'function', we end up with
some functions tagged with the invalid flags coming from other
functions.

---- func: <function test_admin_whitelist_subnets at 0x7fa390a89750>, id: 140340483364688, flags: ('slow',), func_flags: ['slow'], fl set(['slow'])
---- func: <function test_user_whitelist_subnets at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_58918',), func_flags: ['slow', 'redmine_58918'], fl set(['slow'])
---- func: <function test_avp_operations at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_59721', 'redmine_69103'), func_flags: ['slow', 'redmine_58918', 'redmine_59721', 'redmine_69103'], fl set(['slow'])
---- func: <function test_incremental_inheritance at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_70303',), func_flags: ['slow', 'redmine_58918', 'redmine_59721', 'redmine_69103', 'redmine_70303'], fl set(['slow'])
---- func: <function test_incremental_restart at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_60311', 'slow'), func_flags: ['slow', 'redmine_58918', 'redmine_59721', 'redmine_69103', 'redmine_70303', 'redmine_60311', 'slow'], fl set(['slow'])
---- func: <function test_worflow_authentication at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_60536',), func_flags: ['slow', 'redmine_58918', 'redmine_59721', 'redmine_69103', 'redmine_70303', 'redmine_60311', 'slow', 'redmine_60536'], fl set(['slow'])
---- func: <function test_password_change at 0x7fa390a89750>, id: 140340483364688, flags: ('redmine_75002', 'redmine_82440'), func_flags: ['slow', 'redmine_58918', 'redmine_59721', 'redmine_69103', 'redmine_70303', 'redmine_60311', 'slow', 'redmine_60536', 'redmine_75002', 'redmine_82440'], fl set(['slow'])

The solution is not to use the id of the function value as key, but use
the function itself. A function, or a reference to the function, is
hashable and can be compared transparently. So it can be used directly
as the key of the dictionary, and has the correct behaviour.

Change-Id: Ifb3c50fdd015a1540a705c06867624d40715a0fc