wscript/llvm: do not have a maximum llvm supported version.
Having a maximum llvm supported version means that each time a new
llvm/clang version is out, we need to do a patch to support it, even
when nothing actually change in the clang-rewrite-blocks.

This is the case for llvm/clang 16, where it is fully compatible with
our existing code.

Remove the maximum limit so, if we have nothing to do when a new version
is out, this is a huge win, and if not, we still have to make a patch
anyway.

Change-Id: I3ec87fe76a2255facef3d68d70a9b3b86129f7e1