iop/rpc-channel: add the concept of user-version
A custom user version (unsigned 32 bits integer) can now be set in the
IChannel module which will be exchanged during the connection handshake.
The user version is provided with a callback that will be used to check
whether a remote version is accepted or not during the handshake phase
of any connection establishment.

Refs: #98109
Change-Id: I3805b5044a30eaf38306d20cc80dde90fe2fed94