net/https: add SSL callback to log SSL keys to decode exchanges.
Exchanges in HTTPS are encoded with keys. Without them, we can not
read them, with Wireshark for example.
Add a callback in order to log all SSL keys used to decode them.
To do that, it is necessary to call SSL_CTX_set_keylog_callback
method. This method is available only from 1.1.1 SSL version !

Closes: #97561
Change-Id: Ibd7a3fa689a88731b10bd1f5eddabfba8e78cf25