net/https: use SSL_CTX_use_certificate_chain_file to load certificates.
Use SSL_CTX_use_certificate_chain_file instead of
SSL_CTX_use_certificate in order to give possibility to the user to use
a chained certificate. If a standalone certificate is used, the
behavior will be almost the same than previously. The difference is
that SSL will try to load certificate parents from default paths.

Refs: #97323, #98319
Change-Id: I6f642ba7a9b3d4c34455489d5408b5acdc391e71