git describe: remove check for "first-parent" argument
In order to make sure to be able to perform a "git describe
--first-parent" command, which is not available on very old git versions,
the git_describe function was searching for the term "first-parent" in
the help text of the "git describe" command, and was using the result
of this search as an argument for the "git describe" command.

This works on older git versions, but on newer ones (at least with git
2.43.0), the result of this search is not a valid argument, since the
help presents the argument with an optional "no" prefix:
"--[no-]first-parent".

Since there is no build machines with a version of git older than 1.8
anymore, we remove this check, to make sure it will not fail on recent
build machines.

Change-Id: I4e554c32e68aec9b99b415010de92e6ebf367ae9